requirements.txt
streamlit
phi-agent
groq
python-dotenv
